Extract host I/O abstractions from garbage collector
Refactored the `duke-gc` module to fix a Blob anti-pattern where the
`Heap` struct and `lib.rs` file contained complex generational garbage
collection logic intertwined with lower-level host OS file descriptor,
socket, and child process management. Extracted these host I/O management
methods and structs into a dedicated `host.rs` module.
